python/function/trigger_activation/requirements.txt
functions-framework==3.3.0
google-cloud-dataflow-client==0.8.2